+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `CipherSuiteString` returns string for tls.TLS_* constants.
+- `TLSVersionString` returns string for tls.Version* constants.
 
 ## [1.0.1] - 2016-08-28
 ### Added